requirements.txt
spotipy
typer
rich
pandas
python-dotenv
requests
natsort